+++ package.json
@@ -20,7 +20,7 @@
     "react-helmet": "^6.1.0",
     "react-leaflet": "^4.2.0",
     "react-map-gl": "^7.1.3",
-    "react-router-dom": "^6.15.0",
+    "react-router-dom": "^6.24.0",
     "react-scripts": "5.0.1",
     "react-toastify": "^9.1.3",
     "strava-v3": "^2.2.0",
