+++ src/index.js
@@ -6,18 +6,17 @@ import { BrowserRouter, Routes, Route } from "react-router-dom";
 
 import App from "./components/App";
 import "./index.css";
-import About from "./pages/About/About";
-import NotFound from "./pages/NotFound/NotFound";
+import About from "./pages/About";
+import NotFound from "./pages/NotFound";
 
 const root = ReactDOM.createRoot(document.getElementById("root"));
 
 root.render(
   <React.StrictMode>
-    <BrowserRouter basename={process.env.PUBLIC_URL}>
+    <BrowserRouter basename="">
       <Routes>
         <Route path="/" element={<App />} />
-        {/* * should be the last path in the list */}
-        <Route path="about" element={<About />} />
+        <Route path="/about" element={<About />} />
         <Route path="*" element={<NotFound />} />
       </Routes>
     </BrowserRouter>
