--- src/pages/About/About.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-function About() {
-  return (
-    <div>
-      <h1>About Page!</h1>
-      <a href="/"> go home</a>
-    </div>
-  );
-}
-
-export default About;
