--- src/pages/NotFound/NotFound.js
@@ -1,12 +1,0 @@
-import React from "react";
-
-function NotFound() {
-  return (
-    <div>
-      <h1>Page not found!</h1>
-      <a href="/"> go home</a>
-    </div>
-  );
-}
-
-export default NotFound;
